+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.2.0 - 2022-10-03
+
 * Improvement about the filtering by polygon (use a QGIS expression when possible)
 * Add a new option to use the centroid for the filtering by polygon. Available in the 3.10.0 version of desktop plugin.
 * Fix an issue when fetching information from `metadata.txt` in QGIS Server plugins